Added controllers

+++ flix-core/src/main/java/com/epam/mentoring/flixcore/repository/MovieRepository.java
@@ -3,5 +3,5 @@ package com.epam.mentoring.flixcore.repository;
 import com.epam.mentoring.flixcore.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer> {
+public interface MovieRepository extends JpaRepository<Movie, Long> {
 }
